refactor: move config to distinct file
